% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animeR-package.R
\docType{package}
\name{animeR-package}
\alias{animeR}
\alias{animeR-package}
\title{animeR: Explore the world of anime}
\description{
The animeR package is a versatile R tool designed for data exploration and analysis, offering access to two built-in datasets: anime and IKEA furniture. It provides a pre-built app, run using the `run_app()` function, for interactive exploration of IKEA furniture data. Additionally, the package includes two essential functions: `five_stat()`, which computes summary statistics for numeric variables, aiding in data distribution analysis, and `what_ave()`, facilitating the calculation of averages for unique values within categorical variables, enabling data comparison and analysis across categories. With these features, animeR empowers users to unlock valuable insights from their datasets effortlessly.
}
\author{
\strong{Maintainer}: Hoang Do \email{hoangdo@email.com}

}
\keyword{internal}
