% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_stat.R
\name{five_stat}
\alias{five_stat}
\title{Five Statistics}
\usage{
five_stat(input_vector)
}
\arguments{
\item{input_vector}{A vector of data to be summarized.}
}
\value{
A data frame with summaries or frequencies, or NULL.
}
\description{
This function summarizes a vector of data based on its type and length.
If the input is a numeric vector with at least 5 values, it returns a data frame
with the five-number summary. If the input is a character or factor vector,
it returns a data frame with value frequencies. For other types of input or
numeric vectors with less than 5 values, it returns NULL.
}
\examples{
# Numeric vector with at least 5 values
numeric_data <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
five_stat(numeric_data)

# Character vector
character_data <- c("A", "B", "A", "C", "B", "D")
five_stat(character_data)

# Numeric vector with less than 5 values
short_numeric_data <- c(1, 2, 3)
five_stat(short_numeric_data)

# Other types of input
other_data <- c(TRUE, FALSE, TRUE)
five_stat(other_data)

}
