% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_ave.R
\name{what_ave}
\alias{what_ave}
\title{Calculate Average Value by Categorical Variable}
\usage{
what_ave(dataset, categorical_var, numeric_var)
}
\arguments{
\item{dataset}{A data frame containing the dataset.}

\item{categorical_var}{The name of the categorical variable.}

\item{numeric_var}{The name of the rating variable.}
}
\value{
A data frame with unique values of the chosen categorical variable
and their corresponding average numeric values.
}
\description{
This function calculates the average of values from the numeric variable provided
for each unique value within a chosen
categorical variable in the given dataset, even if values are grouped together.
}
\examples{
data <- data.frame(Genre = c("action, sci-fi", "comedy", "action, drama", "sci-fi, fantasy"),
                  Rating = c(4.5, 3.8, 4.0, 3.2))
result <- what_ave(data, categorical_var = "Genre", numeric_var = "Rating")
print(result)

}
